backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
python-dotenv==1.2.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.14
starlette==0.27.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.24.0
uvloop==0.22.1 ; sys_platform != 'win32'
watchfiles==1.1.1
websockets==15.0.1
